Updated sizes in type definitions
Mostly from digging through the go code in various places.
Rust doesn't really like 'usize' and 'isize' much because they're
specifically the size of array indexes; kinda like size_t in C.
So it can be annoying to use them on different platforms or in cross-
platform API's or suck.

+++ ipfs-api/src/client.rs
@@ -1712,7 +1712,7 @@ impl IpfsClient {
     pub fn ping(
         &self,
         peer: &str,
-        count: &Option<usize>,
+        count: &Option<i32>,
     ) -> AsyncStreamResponse<response::PingResponse> {
         self.request_stream_json(&request::Ping { peer, count }, None)
     }
